+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 
 ## 0.0.4
 - Fix defect
+
+## 0.0.5
+- Fix defect
+- New Check internet
+- Add custom host
+- deprecate isHasInternet